simplify trigger form
