+++ src/components/SignalCard/SignalCard.js
@@ -18,7 +18,8 @@ const SignalCard = ({
   goToSignalSettings,
   toggleSignal,
   isUserTheAuthor,
-  deleteEnabled = true
+  deleteEnabled = true,
+  showMoreActions = true
 }) => {
   const isAwaiting = +id <= 0
   const { title, description = '', isPublic, settings: { type } = {} } = signal
@@ -43,7 +44,7 @@ const SignalCard = ({
       </div>
       <div className={styles.wrapper__right}>
         <div onClick={goToSignalSettings}>
-          <div className={styles.upper}>
+          <div className={goToSignalSettings && styles.pointer}>
             <h2 className={styles.title}>{title}</h2>
             {description && (
               <h3 className={styles.description}>
@@ -59,6 +60,7 @@ const SignalCard = ({
         <SignalCardBottom
           signalId={id}
           signal={signal}
+          showMoreActions={showMoreActions}
           removeSignal={removeSignal}
           toggleSignal={toggleSignal}
           isAwaiting={isAwaiting}
@@ -84,12 +86,13 @@ const SignalCardBottom = ({
   isAwaiting = false,
   toggleSignal,
   isUserTheAuthor = true,
-  deleteEnabled
+  deleteEnabled,
+  showMoreActions
 }) => {
   const { isActive, isPublic, title } = signal
   return (
     <div className={styles.bottom}>
-      <DesktopOnly>
+      {showMoreActions && <DesktopOnly>
         <MoreSignalActions
           isUserTheAuthor={isUserTheAuthor}
           removeSignal={removeSignal}
@@ -98,7 +101,7 @@ const SignalCardBottom = ({
           isPublic={isPublic}
           deleteEnabled={deleteEnabled}
         />
-      </DesktopOnly>
+      </DesktopOnly>}
 
       {isPublished ? (
         <h4 className={styles.author}>
@@ -115,7 +118,7 @@ const SignalCardBottom = ({
       ) : (
         <UnpublishedMsg />
       )}
-      {isUserTheAuthor && (
+      {isUserTheAuthor && toggleSignal && (
         <div className={styles.right}>
           <Toggle onClick={toggleSignal} isActive={isActive} />
         </div>
