+++ src/components/SignalCard/SignalCard.module.scss
@@ -57,7 +57,7 @@
   margin-top: 10px;
 }
 
-.upper {
+.pointer {
   cursor: pointer;
 }
