+++ src/ducks/Signals/signalFormManager/signalMaster/SignalMaster.js
@@ -5,11 +5,48 @@ import { compose } from 'recompose'
 import { connect } from 'react-redux'
 import { createTrigger, updateTrigger } from '../../common/actions'
 import Message from '@santiment-network/ui/Message'
-import { mapTriggerToFormProps, mapFormPropsToTrigger } from '../../utils/utils'
+import {mapTriggerToFormProps, mapFormPropsToTrigger, getNewTitle, getNewDescription} from '../../utils/utils'
 import { SIGNAL_ROUTES } from '../../common/constants'
 import TriggerForm from '../signalCrudForm/signal/TriggerForm'
-import styles from '../signalCrudForm/signal/TriggerForm.module.scss'
 import SharedTriggerForm from '../sharedForm/SharedTriggerForm'
+import {DEFAULT_FORM_META_SETTINGS, METRIC_DEFAULT_VALUES, PRICE_PERCENT_CHANGE} from "../../utils/constants";
+import styles from '../signalCrudForm/signal/TriggerForm.module.scss'
+
+const mapFormSettings = (baseSettings, meta) => {
+  const metaFormSettings = { ...DEFAULT_FORM_META_SETTINGS, ...meta }
+
+  const formMetric =
+    metaFormSettings && metaFormSettings.metric
+      ? metaFormSettings.metric.value.value
+      : PRICE_PERCENT_CHANGE
+  let settings = {
+    ...METRIC_DEFAULT_VALUES[formMetric],
+    target: metaFormSettings.target.value
+      ? metaFormSettings.target.value
+      : baseSettings.target,
+    metric: metaFormSettings.metric.value
+      ? metaFormSettings.metric.value
+      : baseSettings.metric,
+    type: metaFormSettings.type.value
+      ? metaFormSettings.type.value
+      : baseSettings.type,
+    signalType: metaFormSettings.signalType.value
+      ? metaFormSettings.signalType.value
+      : baseSettings.signalType,
+    ethAddress: metaFormSettings.ethAddress,
+    ...baseSettings
+  }
+
+  if (!settings.title && !settings.description) {
+    settings = {
+      title: getNewTitle(settings),
+      description: getNewDescription(settings),
+      ...settings
+    }
+  }
+
+  return [settings, metaFormSettings]
+}
 
 const SignalMaster = ({
   canRedirect = true,
@@ -43,8 +80,6 @@ const SignalMaster = ({
     isActive: true,
     isPublic: false
   })
-
-  const triggerSettingsFormData = mapTriggerToFormProps(stateTrigger)
 
   useEffect(
     () => {
@@ -80,6 +115,13 @@ const SignalMaster = ({
 
   const close = onClose || redirect
 
+  const [settings, metaForm] = mapFormSettings(
+    mapTriggerToFormProps(stateTrigger),
+    metaFormSettings
+  )
+
+  debugger
+
   return (
     <div className={styles.wrapper}>
       {!openSharedForm && (
@@ -87,8 +129,8 @@ const SignalMaster = ({
           setTitle={setTitle}
           id={stateTrigger.id}
           isShared={isShared}
-          metaFormSettings={metaFormSettings}
-          settings={triggerSettingsFormData}
+          metaFormSettings={metaForm}
+          settings={settings}
           onSettingsChange={handleSettingsChange}
           onRemovedSignal={close}
           formChangedCallback={formChangedCallback}
@@ -97,11 +139,11 @@ const SignalMaster = ({
 
       {openSharedForm && (
         <SharedTriggerForm
+          id={stateTrigger.id}
           trigger={stateTrigger}
           onOpen={setOpenSharedForm}
           onCreate={handleSettingsChange}
-          metaFormSettings={metaFormSettings}
-          settings={triggerSettingsFormData}
+          settings={settings}
         />
       )}
     </div>
